+++ pkgs/firehose/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.31
 
 - Add PR Health checks for breaking changes.
+- Add PR Health checks for `DO_NOT${'_'}SUBMIT` strings.
 
 ## 0.3.30